requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.24.27
botocore==1.27.75
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.2
django-ninja==1.4.0
django-storages==1.14.5
django-stubs==5.1.3
django-stubs-ext==5.1.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
kombu==5.5.0
multidict==6.2.0
mypy==1.15.0
mypy-extensions==1.0.0
openai==1.66.3
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
s3transfer==0.6.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.18.3
